MULE-10936: Allow the SchedulerService to create pools per app rather that per container - split scheduler service
